refactor(decompose): properly split capture/run into multiple files
