+++ packages/libraries/graph-tools/src/commands/capture/run/implementation.ts
@@ -1,15 +1,11 @@
 import fs from 'node:fs/promises'
 import path from 'node:path'
-import { hydrateCommand, serializeState, type Delta, type State } from '@vt/graph-state'
-import { computeDrift, type DriftData, type FsContentById } from '@vt/graph-tools/debug/state/drift'
+import { hydrateCommand, type Delta } from '@vt/graph-state'
 import { normalizeChord } from '@vt/graph-tools/debug/input/normalizeChord'
 import { pressChord } from '@vt/graph-tools/debug/input/pressChord'
+import { err, ok, type Response } from '@vt/graph-tools/debug/protocol/Response'
 import { openDebugSession, type PageLike as SessionPageLike } from '@vt/graph-tools/debug/protocol/playwrightSession'
 import { resolveDebugInstance } from '@vt/graph-tools/debug/protocol/portResolution'
-import { projectStateToCyDump } from '@vt/graph-tools/debug/state/projectedCyDump'
-import { err, ok } from '@vt/graph-tools/debug/protocol/Response'
-import type { Response } from '@vt/graph-tools/debug/protocol/Response'
-import { parseCyDump, type CyDump } from '@vt/graph-tools/debug/state/cyStateShape'
 import {
   STEP_SPEC_SELECTOR_NOTE,
   validateStepSpec,
@@ -22,15 +18,11 @@ import {
   readFlagValue,
   type DebugTargetArgs,
 } from '../../core/argv'
-
-const DEFAULT_RUN_DIR = '/tmp/vt-debug/run'
-const DEFAULT_WAIT_FOR_TIMEOUT_MS = 5000
-const DEFAULT_CLICK_TIMEOUT_MS = 2000
+import { runObservations } from './observations'
 
 type ResolvedDebugTarget = Awaited<ReturnType<typeof resolveDebugInstance>>
 type DebugInstance = Extract<ResolvedDebugTarget, { ok: true }>['instance']
-
-interface PageLike extends SessionPageLike {
+type PageLike = SessionPageLike & {
   click(selector: string, options?: { timeout?: number }): Promise<void>
   goto(url: string): Promise<unknown>
   mouse: {
@@ -42,30 +34,6 @@ interface PageLike extends SessionPageLike {
     options?: { timeout?: number },
   ): Promise<unknown>
 }
-
-type FloatingEditorRect = {
-  id: string
-  x: number
-  y: number
-  w: number
-  h: number
-  intersectsViewport: boolean
-}
-
-type DomProbes = {
-  cyNodeCount: number
-  floatingEditors: string[]
-  floatingEditorRects: FloatingEditorRect[]
-  selectedNodeHasEditor: boolean
-  rootClientHeightPx: number
-  sidebarWrapperVisible: boolean
-}
-
-type TapNodePlan =
-  | { ok: true; strategy: 'mouse'; x: number; y: number }
-  | { ok: true; strategy: 'emit' }
-  | { ok: false; error: string }
-
 type RunOptions = {
   specSource: string
   screenshotEach: boolean
@@ -78,12 +46,10 @@ type RunOptions = {
   pid?: number
   vault?: string
 }
-
 type LoadedSteps = {
   steps: StepSpec[]
   source: string
 }
-
 type RunStepOutput = {
   step: StepSpec
   ok: boolean
@@ -94,7 +60,6 @@ type RunStepOutput = {
   drift?: string
   state?: string
 }
-
 type RunResult = {
   source: string
   bundle: {
@@ -103,136 +68,14 @@ type RunResult = {
     outputs: RunStepOutput[]
   }
 }
+type StateCaptureOverlay = ReturnType<typeof runObservations.createStateCaptureOverlay>
 
-type SerializedStateSnapshot = ReturnType<typeof serializeState>
+const DEFAULT_RUN_DIR = '/tmp/vt-debug/run'
+const DEFAULT_WAIT_FOR_TIMEOUT_MS = 5000
+const DEFAULT_CLICK_TIMEOUT_MS = 2000
 
-type StateCaptureOverlay = {
-  collapseSet: string[]
-  selection: string[]
-  rootsLoaded: string[]
-  layout: {
-    zoom?: number
-    pan?: { x: number; y: number }
-  }
-}
-
-function sortStrings(values: readonly string[]): string[] {
-  return [...values].sort((left, right) => left.localeCompare(right))
-}
-
-function removeValues(values: readonly string[], removed: readonly string[] | undefined): string[] {
-  if (!removed || removed.length === 0) return [...values]
-  const removedSet = new Set(removed)
-  return values.filter(value => !removedSet.has(value))
-}
-
-function appendUnique(values: readonly string[], added: readonly string[] | undefined): string[] {
-  if (!added || added.length === 0) return [...values]
-
-  const next = [...values]
-  const seen = new Set(next)
-  for (const value of added) {
-    if (seen.has(value)) continue
-    next.push(value)
-    seen.add(value)
-  }
-  return next
-}
-
-function updateOrderedValues(
-  values: readonly string[],
-  removed: readonly string[] | undefined,
-  added: readonly string[] | undefined,
-): string[] {
-  return appendUnique(removeValues(values, removed), added)
-}
-
-function createStateCaptureOverlay(state: State): StateCaptureOverlay {
-  return {
-    collapseSet: sortStrings([...state.collapseSet]),
-    selection: [...state.selection],
-    rootsLoaded: sortStrings([...state.roots.loaded]),
-    layout: {
-      ...(state.layout.zoom !== undefined ? { zoom: state.layout.zoom } : {}),
-      ...(state.layout.pan !== undefined ? { pan: state.layout.pan } : {}),
-    },
-  }
-}
-
-function applyDeltaToStateCaptureOverlay(
-  overlay: StateCaptureOverlay,
-  delta: Delta | null | undefined,
-): StateCaptureOverlay {
-  if (!delta) return overlay
-
-  const next: StateCaptureOverlay = {
-    collapseSet: updateOrderedValues(
-      overlay.collapseSet,
-      delta.collapseRemoved,
-      delta.collapseAdded,
-    ),
-    selection: updateOrderedValues(
-      overlay.selection,
-      delta.selectionRemoved,
-      delta.selectionAdded,
-    ),
-    rootsLoaded: updateOrderedValues(
-      overlay.rootsLoaded,
-      delta.rootsUnloaded,
-      delta.rootsLoaded,
-    ),
-    layout: {
-      ...overlay.layout,
-      ...(delta.layoutChanged?.zoom !== undefined ? { zoom: delta.layoutChanged.zoom } : {}),
-      ...(delta.layoutChanged?.pan !== undefined ? { pan: delta.layoutChanged.pan } : {}),
-    },
-  }
-
-  return {
-    ...next,
-    collapseSet: sortStrings(next.collapseSet),
-    rootsLoaded: sortStrings(next.rootsLoaded),
-  }
-}
-
-function buildCapturedSerializedState(
-  state: State,
-  overlay: StateCaptureOverlay | null | undefined,
-  rendered: CyDump | null,
-): SerializedStateSnapshot {
-  const serialized = serializeState(state)
-  const overlayZoom = overlay?.layout.zoom
-  const overlayPan = overlay?.layout.pan
-
-  const layout = {
-    ...serialized.layout,
-    ...(overlayZoom !== undefined ? { zoom: overlayZoom } : {}),
-    ...(overlayPan !== undefined ? { pan: overlayPan } : {}),
-    ...(serialized.layout.zoom === undefined && overlayZoom === undefined && rendered
-      ? { zoom: rendered.viewport.zoom }
-      : {}),
-    ...(serialized.layout.pan === undefined && overlayPan === undefined && rendered
-      ? { pan: rendered.viewport.pan }
-      : {}),
-  }
-
-  if (!overlay) {
-    return {
-      ...serialized,
-      layout,
-    }
-  }
-
-  return {
-    ...serialized,
-    roots: {
-      ...serialized.roots,
-      loaded: [...overlay.rootsLoaded],
-    },
-    collapseSet: [...overlay.collapseSet],
-    selection: [...overlay.selection],
-    layout,
-  }
+function isErrorResponse<T>(value: T | Response<never>): value is Response<never> {
+  return typeof value === 'object' && value !== null && 'ok' in value && value.ok === false
 }
 
 function usage(message?: string): Response<never> {
@@ -265,41 +108,27 @@ function parseRunArgs(argv: string[]): RunOptions | Response<never> {
   try {
     for (let i = 0; i < argv.length; i += 1) {
       const arg = argv[i]
-      if (arg === '--screenshot-each') {
-        screenshotEach = true
-      } else if (arg === '--console-each') {
-        consoleEach = true
-      } else if (arg === '--drift-each') {
-        driftEach = true
-      } else if (arg === '--state-each') {
-        stateEach = true
-      } else if (arg === '--stop-on-error') {
-        stopOnError = parseBooleanFlag('--stop-on-error', argv[++i])
-      } else if (arg.startsWith('--stop-on-error=')) {
+      if (arg === '--screenshot-each') screenshotEach = true
+      else if (arg === '--console-each') consoleEach = true
+      else if (arg === '--drift-each') driftEach = true
+      else if (arg === '--state-each') stateEach = true
+      else if (arg === '--stop-on-error') stopOnError = parseBooleanFlag('--stop-on-error', argv[++i])
+      else if (arg.startsWith('--stop-on-error=')) {
         stopOnError = parseBooleanFlag('--stop-on-error', arg.slice('--stop-on-error='.length))
-      } else if (arg === '--out') {
-        outDir = path.resolve(readFlagValue('--out', argv[++i]))
-      } else if (arg.startsWith('--out=')) {
-        outDir = path.resolve(readFlagValue('--out', arg.slice('--out='.length)))
-      } else if (arg.startsWith('--')) {
+      } else if (arg === '--out') outDir = path.resolve(readFlagValue('--out', argv[++i]))
+      else if (arg.startsWith('--out=')) outDir = path.resolve(readFlagValue('--out', arg.slice('--out='.length)))
+      else if (arg.startsWith('--')) {
         const debugTargetFlag = consumeDebugTargetFlag(argv, i, target)
-        if (!debugTargetFlag.matched) {
-          return usage(`unknown argument: ${arg}`)
-        }
+        if (!debugTargetFlag.matched) return usage(`unknown argument: ${arg}`)
         i = debugTargetFlag.nextIndex
-      } else if (specSource === undefined) {
-        specSource = arg
-      } else {
-        return usage(`unexpected positional argument: ${arg}`)
-      }
+      } else if (specSource === undefined) specSource = arg
+      else return usage(`unexpected positional argument: ${arg}`)
     }
   } catch (e) {
     return usage(String(e))
   }
 
-  if (!specSource) {
-    return usage('run requires a spec file path or inline JSON array')
-  }
+  if (!specSource) return usage('run requires a spec file path or inline JSON array')
 
   return {
     specSource,
@@ -313,10 +142,6 @@ function parseRunArgs(argv: string[]): RunOptions | Response<never> {
     pid: target.pid,
     vault: target.vault,
   }
-}
-
-function isErrorResponse<T>(value: T | Response<never>): value is Response<never> {
-  return typeof value === 'object' && value !== null && 'ok' in value && value.ok === false
 }
 
 function isRecord(input: unknown): input is Record<string, unknown> {
@@ -337,16 +162,12 @@ function parseStepSpecsDocument(raw: string): LoadedSteps | Response<never> {
       ? parsed.steps
       : null
 
-  if (!rawSteps) {
-    return usage('spec must be a JSON array or an object with a steps array')
-  }
+  if (!rawSteps) return usage('spec must be a JSON array or an object with a steps array')
 
   const steps: StepSpec[] = []
   for (let i = 0; i < rawSteps.length; i += 1) {
     const result = validateStepSpec(rawSteps[i])
-    if (!result.ok) {
-      return usage(`invalid step at index ${i}: ${result.error}`)
-    }
+    if (!result.ok) return usage(`invalid step at index ${i}: ${result.error}`)
     steps.push(result.step)
   }
 
@@ -370,14 +191,6 @@ async function loadSteps(specSource: string): Promise<LoadedSteps | Response<nev
   } catch (e) {
     return usage(`unable to read spec file: ${String(e)}`)
   }
-}
-
-async function resolveTarget(options: RunOptions) {
-  const pick = await resolveDebugInstance({ port: options.port, pid: options.pid, vault: options.vault })
-  if (!pick.ok) {
-    return { ok: false as const, response: err('run', pick.message, pick.hint, 2) }
-  }
-  return { ok: true as const, instance: pick.instance }
 }
 
 async function focusTarget(page: SessionPageLike, selector: string): Promise<void> {
@@ -411,110 +224,22 @@ async function focusTarget(page: SessionPageLike, selector: string): Promise<voi
     }
   })()`)
 
-  if (!result.ok) {
-    throw new Error(result.error)
-  }
-}
-
-async function executeStep(
-  page: PageLike,
-  step: StepSpec,
-  instance: DebugInstance,
-): Promise<Delta | null> {
-  if ('dispatch' in step) {
-    const transport = createLiveTransport(instance.mcpPort)
-    return transport.dispatchLiveCommand(hydrateCommand(step.dispatch))
-  }
-
-  if ('click' in step) {
-    await page.click(step.click, { timeout: DEFAULT_CLICK_TIMEOUT_MS })
-    return null
-  }
-
-  if ('tapNode' in step) {
-    await executeTapNodeStep(page, step.tapNode)
-    return null
-  }
-
-  if ('type' in step) {
-    if (step.selector) {
-      await focusTarget(page, step.selector)
-    }
-    await page.keyboard.type(step.type)
-    return null
-  }
-
-  if ('press' in step) {
-    if (step.selector) {
-      await focusTarget(page, step.selector)
-    }
-    await pressChord(page, normalizeChord(step.press))
-    return null
-  }
-
-  if ('wait' in step) {
-    await new Promise(resolve => setTimeout(resolve, step.wait))
-    return null
-  }
-
-  if ('waitFor' in step) {
-    await page.waitForSelector(step.waitFor, {
-      timeout: step.timeoutMs ?? DEFAULT_WAIT_FOR_TIMEOUT_MS,
-    })
-    return null
-  }
-
-  await page.goto(step.navigate)
-  return null
-}
-
-async function fetchRendered(page: SessionPageLike): Promise<CyDump> {
-  const raw = await page.evaluate(() => {
-    const helper = (window as unknown as Record<string, unknown>)['__vtDebug__']
-    if (!helper) return null
-    const cy = (helper as Record<string, unknown>)['cy']
-    return typeof cy === 'function' ? (cy as () => unknown)() : null
-  })
-
-  if (raw === null) {
-    throw new Error('window.__vtDebug__.cy() unavailable')
-  }
-
-  return parseCyDump(raw)
-}
-
-async function tryFetchRendered(page: SessionPageLike): Promise<CyDump | null> {
-  try {
-    return await fetchRendered(page)
-  } catch {
-    return null
-  }
-}
-
-async function snapshotFsContent(state: State): Promise<FsContentById> {
-  const entries = await Promise.all(
-    Object.keys(state.graph.nodes).map(async (nodeId) => {
-      try {
-        const content = await fs.readFile(nodeId, 'utf8')
-        return [nodeId, content] as const
-      } catch {
-        return [nodeId, null] as const
-      }
-    }),
-  )
-  return Object.fromEntries(entries)
-}
-
-async function writeJsonFile(filePath: string, data: unknown): Promise<void> {
-  await fs.writeFile(filePath, `${JSON.stringify(data, null, 2)}\n`, 'utf8')
+  if (!result.ok) throw new Error(result.error)
 }
 
 async function sleep(ms: number): Promise<void> {
   await new Promise(resolve => setTimeout(resolve, ms))
 }
 
-async function planTapNode(page: SessionPageLike, nodeId: string): Promise<TapNodePlan> {
-  return page.evaluate<TapNodePlan>(String.raw`(() => {
+async function planTapNode(
+  page: SessionPageLike,
+  nodeId: string,
+): Promise<
+  | { ok: true; strategy: 'mouse'; x: number; y: number }
+  | { ok: true; strategy: 'emit' }
+  | { ok: false; error: string }
+> {
+  return page.evaluate(String.raw`(() => {
     const targetId = ${JSON.stringify(nodeId)}
     const cy = window.cytoscapeInstance
     if (!cy) {
@@ -559,159 +284,73 @@ async function emitTapOnNode(page: SessionPageLike, nodeId: string): Promise<voi
     return { ok: true }
   })()`)
 
-  if (!result.ok) {
-    throw new Error(result.error ?? `tapNode fallback failed for ${nodeId}`)
-  }
+  if (!result.ok) throw new Error(result.error ?? `tapNode fallback failed for ${nodeId}`)
 }
 
 async function executeTapNodeStep(page: PageLike, nodeId: string): Promise<void> {
   const plan = await planTapNode(page, nodeId)
-  if (!plan.ok) {
-    throw new Error(plan.error)
-  }
+  if (!plan.ok) throw new Error(plan.error)
 
-  if (plan.strategy === 'mouse') {
-    await page.mouse.click(plan.x, plan.y)
-  } else {
-    await emitTapOnNode(page, nodeId)
-  }
+  if (plan.strategy === 'mouse') await page.mouse.click(plan.x, plan.y)
+  else await emitTapOnNode(page, nodeId)
 
   await sleep(400)
 }
 
-async function captureScreenshot(page: PageLike, filePath: string): Promise<string> {
-  await page.screenshot({ path: filePath, type: 'png', fullPage: true })
-  return filePath
-}
-
-async function captureJsonObservation(
-  filePath: string,
-  producer: () => Promise<unknown>,
-): Promise<{ path: string; error?: string }> {
-  try {
-    const data = await producer()
-    await writeJsonFile(filePath, data)
-    return { path: filePath }
-  } catch (e) {
-    const error = String(e)
-    await writeJsonFile(filePath, { ok: false, error })
-    return { path: filePath, error }
-  }
-}
-
-async function captureConsoleTail(page: SessionPageLike): Promise<unknown> {
-  const consoleTail = await page.evaluate(() => {
-    const helper = (window as unknown as Record<string, unknown>)['__vtDebug__']
-    if (!helper) return null
-    const readConsole = (helper as Record<string, unknown>)['console']
-    return typeof readConsole === 'function' ? (readConsole as (...args: unknown[]) => unknown)(500) : null
-  })
-
-  if (consoleTail === null) {
-    throw new Error('window.__vtDebug__.console() unavailable')
+async function executeStep(
+  page: PageLike,
+  step: StepSpec,
+  instance: Pick<DebugInstance, 'mcpPort'>,
+): Promise<Delta | null> {
+  if ('dispatch' in step) {
+    const transport = createLiveTransport(instance.mcpPort)
+    return transport.dispatchLiveCommand(hydrateCommand(step.dispatch))
   }
 
-  return consoleTail
-}
+  if ('click' in step) {
+    await page.click(step.click, { timeout: DEFAULT_CLICK_TIMEOUT_MS })
+    return null
+  }
 
-async function captureDomProbes(page: SessionPageLike): Promise<DomProbes> {
-  return page.evaluate<DomProbes>(String.raw`(() => {
-    const safe = (value) => (typeof value === 'number' && Number.isFinite(value) ? value : 0)
-    const intersectsViewport = rect =>
-      rect.right > 0 &&
-      rect.bottom > 0 &&
-      rect.left < window.innerWidth &&
-      rect.top < window.innerHeight
-    const isVisible = element => {
-      if (!(element instanceof HTMLElement)) return false
-      const style = window.getComputedStyle(element)
-      const rect = element.getBoundingClientRect()
-      return (
-        style.display !== 'none' &&
-        style.visibility !== 'hidden' &&
-        safe(rect.width) > 0 &&
-        safe(rect.height) > 0
-      )
-    }
+  if ('tapNode' in step) {
+    await executeTapNodeStep(page, step.tapNode)
+    return null
+  }
 
-    const floatingEditorElements = Array.from(
-      document.querySelectorAll('[id^="window-"][id$="-editor"]'),
-    ).filter(element => element instanceof HTMLElement)
+  if ('type' in step) {
+    if (step.selector) await focusTarget(page, step.selector)
+    await page.keyboard.type(step.type)
+    return null
+  }
 
-    const floatingEditorRects = floatingEditorElements.map(element => {
-      const rect = element.getBoundingClientRect()
-      return {
-        id: element.id,
-        x: safe(rect.left),
-        y: safe(rect.top),
-        w: safe(rect.width),
-        h: safe(rect.height),
-        intersectsViewport: intersectsViewport(rect),
-      }
+  if ('press' in step) {
+    if (step.selector) await focusTarget(page, step.selector)
+    await pressChord(page, normalizeChord(step.press))
+    return null
+  }
+
+  if ('wait' in step) {
+    await new Promise(resolve => setTimeout(resolve, step.wait))
+    return null
+  }
+
+  if ('waitFor' in step) {
+    await page.waitForSelector(step.waitFor, {
+      timeout: step.timeoutMs ?? DEFAULT_WAIT_FOR_TIMEOUT_MS,
     })
-
-    const rectById = new Map(floatingEditorRects.map(rect => [rect.id, rect]))
-    const selectedNodeIds =
-      typeof window.cytoscapeInstance?.$ === 'function'
-        ? window.cytoscapeInstance.$(':selected').map(node => node.id())
-        : []
-
-    const selectedNodeHasEditor =
-      selectedNodeIds.length > 0 &&
-      selectedNodeIds.every(nodeId => rectById.get('window-' + nodeId + '-editor')?.intersectsViewport === true)
-
-    const root = document.getElementById('root')
-    const sidebarWrapper = document.querySelector('.sidebar-wrapper')
-    const cyNodeCount =
-      typeof window.cytoscapeInstance?.nodes === 'function'
-        ? window.cytoscapeInstance.nodes().length
-        : 0
-
-    return {
-      cyNodeCount,
-      floatingEditors: floatingEditorElements.map(element => element.id),
-      floatingEditorRects,
-      selectedNodeHasEditor,
-      rootClientHeightPx: root instanceof HTMLElement ? root.clientHeight : 0,
-      sidebarWrapperVisible: isVisible(sidebarWrapper),
-    }
-  })()`)
-}
-
-async function captureSerializedState(
-  instance: DebugInstance,
-  page: SessionPageLike,
-  overlay?: StateCaptureOverlay | null,
-): Promise<unknown> {
-  const transport = createLiveTransport(instance.mcpPort)
-  const [state, domProbes, rendered] = await Promise.all([
-    transport.getLiveState(),
-    captureDomProbes(page),
-    tryFetchRendered(page),
-  ])
-
-  return {
-    ...buildCapturedSerializedState(state, overlay, rendered),
-    domProbes,
+    return null
   }
+
+  await page.goto(step.navigate)
+  return null
 }
 
-async function captureDrift(instance: DebugInstance, page: SessionPageLike): Promise<unknown> {
-  const transport = createLiveTransport(instance.mcpPort)
-  const [state, rendered] = await Promise.all([
-    transport.getLiveState(),
-    fetchRendered(page),
-  ])
-  const projected = projectStateToCyDump(state)
-  const fsContentById = await snapshotFsContent(state)
-  return computeDrift(
-    {
-      ...state,
-      fsContentById,
-    } satisfies DriftData,
-    projected,
-    rendered,
-  )
+async function resolveTarget(options: RunOptions) {
+  const pick = await resolveDebugInstance({ port: options.port, pid: options.pid, vault: options.vault })
+  if (!pick.ok) {
+    return { ok: false as const, response: err('run', pick.message, pick.hint, 2) }
+  }
+  return { ok: true as const, instance: pick.instance }
 }
 
 function stepBaseName(index: number): string {
@@ -730,7 +369,7 @@ async function runSteps(
   if (options.stateEach) {
     try {
       const transport = createLiveTransport(instance.mcpPort)
-      captureOverlay = createStateCaptureOverlay(await transport.getLiveState())
+      captureOverlay = runObservations.createStateCaptureOverlay(await transport.getLiveState())
     } catch {
       captureOverlay = null
     }
@@ -739,71 +378,29 @@ async function runSteps(
   for (let i = 0; i < steps.length; i += 1) {
     const step = steps[i]
     const output: RunStepOutput = { step, ok: true }
-    const observationErrors: string[] = []
     let stepDelta: Delta | null = null
 
     try {
       stepDelta = await executeStep(page, step, instance)
       if (captureOverlay) {
-        captureOverlay = applyDeltaToStateCaptureOverlay(captureOverlay, stepDelta)
+        captureOverlay = runObservations.applyDeltaToStateCaptureOverlay(captureOverlay, stepDelta)
       }
     } catch (e) {
       output.ok = false
       output.error = String(e)
     }
 
-    const baseName = path.join(options.outDir, stepBaseName(i))
-
-    if (options.screenshotEach) {
-      try {
-        output.screenshot = await captureScreenshot(page, `${baseName}.png`)
-      } catch (e) {
-        observationErrors.push(`screenshot: ${String(e)}`)
-      }
-    }
-
-    if (options.consoleEach) {
-      const result = await captureJsonObservation(
-        `${baseName}.console.json`,
-        () => captureConsoleTail(page),
-      )
-      output.console = result.path
-      if (result.error) {
-        observationErrors.push(`console: ${result.error}`)
-      }
-    }
-
-    if (options.driftEach) {
-      const result = await captureJsonObservation(
-        `${baseName}.drift.json`,
-        () => captureDrift(instance, page),
-      )
-      output.drift = result.path
-      if (result.error) {
-        observationErrors.push(`drift: ${result.error}`)
-      }
-    }
-
-    if (options.stateEach) {
-      const result = await captureJsonObservation(
-        `${baseName}.state.json`,
-        () => captureSerializedState(instance, page, captureOverlay),
-      )
-      output.state = result.path
-      if (result.error) {
-        observationErrors.push(`state: ${result.error}`)
-      }
-    }
-
-    if (observationErrors.length > 0) {
-      output.observationErrors = observationErrors
-    }
+    Object.assign(output, await runObservations.captureStepObservations({
+      page,
+      baseName: path.join(options.outDir, stepBaseName(i)),
+      captureOverlay,
+      getLiveState: () => createLiveTransport(instance.mcpPort).getLiveState(),
+      options,
+    }))
 
     outputs.push(output)
 
-    if (!output.ok && options.stopOnError) {
-      break
-    }
+    if (!output.ok && options.stopOnError) break
   }
 
   return outputs
@@ -857,15 +454,13 @@ async function runHandler(argv: string[]): Promise<Response<RunResult>> {
       3,
     )
   } finally {
-    if (session) {
-      await session.close()
-    }
+    if (session) await session.close()
   }
 }
 
 export const runImplementation = {
-  applyDeltaToStateCaptureOverlay,
-  buildCapturedSerializedState,
-  createStateCaptureOverlay,
+  applyDeltaToStateCaptureOverlay: runObservations.applyDeltaToStateCaptureOverlay,
+  buildCapturedSerializedState: runObservations.buildCapturedSerializedState,
+  createStateCaptureOverlay: runObservations.createStateCaptureOverlay,
   runHandler,
 }
