+++ packages/libraries/graph-tools/src/commands/capture/run/index.ts
@@ -1,0 +1,1 @@
+export { runImplementation } from './implementation'
