+++ packages/libraries/graph-tools/src/commands/capture/run/types.ts
@@ -1,0 +1,20 @@
+export type RunTypes = {
+  RunStepOutput: {
+    step: unknown
+    ok: boolean
+    error?: string
+    observationErrors?: string[]
+    screenshot?: string
+    console?: string
+    drift?: string
+    state?: string
+  }
+  RunResult: {
+    source: string
+    bundle: {
+      dir: string
+      stepCount: number
+      outputs: RunTypes['RunStepOutput'][]
+    }
+  }
+}
